% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hyper_params.R
\name{check_hyper_params}
\alias{check_hyper_params}
\title{Check input parameters}
\usage{
check_hyper_params(X_names, params)
}
\arguments{
\item{X_names}{The observed covariates names.}

\item{params}{The list of parameters required to run the function.}
}
\value{
A modified input \code{params}. A list of parameters that might be changed during
the checks.
}
\description{
Checks consistency in input (hyper) parameters for the \code{cre} function.
}
\keyword{internal}
