% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input parameters}
\usage{
check_input(y, z, X, params)
}
\arguments{
\item{params}{TBD}
}
\value{

}
\description{
Checks input parameters for the cre function.
}
