% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_method_params.R
\name{check_method_params}
\alias{check_method_params}
\title{Check input parameters}
\usage{
check_method_params(y, params)
}
\arguments{
\item{y}{The observed response vector.}

\item{params}{The list of parameters required to run the ite functions.}
}
\value{
A list of params that might be changed during checks.
}
\description{
Checks input parameters for the cre function.
}
