% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_anomalous_rules.R
\name{discard_anomalous_rules}
\alias{discard_anomalous_rules}
\title{Discard Anomalous (Causal) Decision Rules}
\usage{
discard_anomalous_rules(rules_matrix, rules_list, t)
}
\arguments{
\item{rules_matrix}{The causal rules matrix.}

\item{rules_list}{The list of the causal rules.}

\item{t}{The threshold to define too generic or too specific rules.}
}
\value{
A list with:
\itemize{
\item the filtered rules matrix
\item the filtered rules list
}
}
\description{
Discard rules with too few or too many observations.
}
\keyword{internal}
