% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_correlated_rules.R
\name{discard_correlated_rules}
\alias{discard_correlated_rules}
\title{Discard Correlated (Causal) Decision Rules}
\usage{
discard_correlated_rules(rules_matrix, rules_list, t)
}
\arguments{
\item{rules_matrix}{The causal rules matrix.}

\item{rules_list}{The list of the causal rules.}

\item{t}{The threshold to define correlated rules.}
}
\value{
A list with:
\itemize{
\item the filtered rules matrix
\item the filtered rules list
}
}
\description{
Discard redundant correlated rules.
}
\keyword{internal}
