% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_causal_rules.R
\name{discover_causal_rules}
\alias{discover_causal_rules}
\title{Discover Causal Decision Rules}
\usage{
discover_causal_rules(
  rules_matrix_std,
  rules_list,
  ite_std,
  stability_selection,
  cutoff,
  pfer
)
}
\arguments{
\item{rules_matrix_std}{The standardized causal rules matrix.}

\item{rules_list}{A vector of causal rules.}

\item{ite_std}{The standardized ITE.}

\item{stability_selection}{Whether or not using stability selection to
select the causal rules.}

\item{cutoff}{Threshold defining the minimum cutoff value for the stability
scores.}

\item{pfer}{Upper bound for the per-family error rate (tolerated amount of
falsely selected rules).}
}
\value{
List of the Causal Decision Rules discovered
}
\description{
Discover Causal Decision Rules by randomization-based tests
}
\keyword{internal}
