% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{Title}
\usage{
estimate_cate(tau, rules_matrix)
}
\arguments{
\item{tau}{input value description}

\item{rules_matrix}{input value description}
}
\value{

}
\description{
description
}
\examples{
TBD
}
