% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite.R
\name{estimate_ite}
\alias{estimate_ite}
\title{Estimate the Individual Treatment Effect (ITE)}
\usage{
estimate_ite(
  y,
  z,
  X,
  ite_method,
  include_ps,
  ps_method,
  oreg_method,
  binary,
  X_names,
  include_offset,
  offset_name
)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The observed treatment vector.}

\item{X}{The covariate matrix.}

\item{ite_method}{the method for estimating the Individual Treatment Effect:
\itemize{
\item \code{ipw}: Inverse Propensity Weighting
\item \code{sipw}: Stabilized Inverse Propensity Weighting
\item \code{aipw}: Augmented Inverse Propensity Weighting
\item \code{oreg}: Outcome Regression
\item \code{bart}: Bayesian Additive Regression Trees
\item \code{xbart}: Accelerated Bayesian Additive Regression Trees
\item \code{bcf}: Bayesian Causal Forest
\item \code{xbcf}: Accelerated Bayesian Causal Forest
\item \code{cf}: Causal Forest
\item \code{poisson}: Poisson Estimation
}}

\item{include_ps}{whether or not to include propensity score estimate as a
covariate in ITE estimation}

\item{ps_method}{estimation method for the propensity score}

\item{oreg_method}{the estimation model for the outcome regressions in estimate_ite_aipw}

\item{binary}{whether or not the outcome is binary}

\item{X_names}{the names of the covariates}

\item{include_offset}{whether or not to include an offset when estimating
the ITE, for poisson only}

\item{offset_name}{the name of the offset, if it is to be included}
}
\value{
A list that includes:
\itemize{
\item raw ITE estimates
\item standardized ITE estimates, and
\item standard deviations for the ITE estimates.
}
}
\description{
Estimates the Individual Treatment Effect given a response vector,
a treatment vector, a covariate matrix, and a desired algorithm.
}
\examples{
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- dataset[["y"]]
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
ite_method <- "bart"
include_ps <- TRUE
ps_method <- "SL.xgboost"
or_method <- NA
binary <- FALSE
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite(y, z, X, ite_method, include_ps, ps_method, or_method,
                         binary, X_names, include_offset, offset_name)

}
