% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_or.R
\name{estimate_ite_or}
\alias{estimate_ite_or}
\title{Title}
\usage{
estimate_ite_or(y, z, X)
}
\arguments{
\item{y}{input value description}

\item{z}{input value description}

\item{X}{input value description}
}
\value{

}
\description{
description
}
\examples{
TBD

}
