% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_oreg.R
\name{estimate_ite_oreg}
\alias{estimate_ite_oreg}
\title{Estimate the Individual Treatment Effect (ITE) using Outcome Regression
(OREG)}
\usage{
estimate_ite_oreg(y, z, X)
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}
}
\value{
A vector of ITE estimates
}
\description{
Estimates the Individual Treatment Effect (ITE) using Outcome Regression
given a response vector, a treatment vector, and a features matrix.
}
\keyword{internal}
