% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_poisson.R
\name{estimate_ite_poisson}
\alias{estimate_ite_poisson}
\title{Estimate the Individual Treatment Effect using Poisson Regression}
\usage{
estimate_ite_poisson(y, z, X, X_names, offset)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The features matrix.}

\item{X_names}{The names of the covariates.}

\item{offset}{Name of the covariate to use as offset (i.e. 'x1') to model the
corresponding outcome rate. NULL to model drectly the outcome counts without
offset.}
}
\value{
A vector of ITE estimates
}
\description{
Estimates the Individual Treatment Effect using Poisson Regression given a
response vector, a treatment vector, and a features matrix.
}
\keyword{internal}
