% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_slearner.R
\name{estimate_ite_slearner}
\alias{estimate_ite_slearner}
\title{Estimate the Individual Treatment Effect (ITE) using S-Learner}
\usage{
estimate_ite_slearner(y, z, X, learner_y = "SL.xgboost")
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{learner_y}{An estimation model for the outcome.}
}
\value{
A list of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using S-Learner given a response
vector, a treatment vector, a features matrix and estimation model for the
outcome.
}
\keyword{internal}
