% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_tpoisson.R
\name{estimate_ite_tpoisson}
\alias{estimate_ite_tpoisson}
\title{Estimate the Individual Treatment Effect (ITE) using T-Poisson Regression}
\usage{
estimate_ite_tpoisson(y, z, X, offset)
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{offset}{A name of the covariate to use as offset (i.e. 'x1') to model
the corresponding outcome rate. \code{NULL} to model directly the outcome counts
without offset.}
}
\value{
A vector of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using Poisson Regression given a
response vector, a treatment vector, and a features matrix.
}
\keyword{internal}
