% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_xbart.R
\name{estimate_ite_xbart}
\alias{estimate_ite_xbart}
\title{Estimate the Individual Treatment Effect using Accelerated Bayesian Additive Regression Trees}
\usage{
estimate_ite_xbart(y, z, X, include_ps, ps_method)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{include_ps}{whether or not to include propensity score estimate as a covariate in ITE estimation}

\item{ps_method}{estimation method for the propensity score}
}
\value{
a list of ITE estimates and standard deviations for the ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Accelerated Bayesian Additive Regression Trees given a response vector, a treatment vector, and a features matrix
}
\examples{
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- dataset[["y"]]
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
include_ps <- TRUE
ps_method <- "SL.xgboost"

ite_list <- estimate_ite_xbart(y, z, X, include_ps, ps_method)

}
