% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effect_modifiers.R
\name{extract_effect_modifiers}
\alias{extract_effect_modifiers}
\title{Extract effect modifiers}
\usage{
extract_effect_modifiers(rules_list, X_names)
}
\arguments{
\item{rules_list}{A list of (causal) decision rules.}

\item{X_names}{A list of the covariate names.}
}
\value{
A list of the effect modifiers.
}
\description{
Extracts the effect modifiers from a list of (causal) decision rules.
}
\keyword{internal}
