% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_correlated_rules.R
\name{filter_correlated_rules}
\alias{filter_correlated_rules}
\title{Filter Correlated Decision Rules}
\usage{
filter_correlated_rules(rules_matrix, rules_list, t_corr)
}
\arguments{
\item{rules_matrix}{The causal rules matrix.}

\item{rules_list}{The list of the causal rules.}

\item{t_corr}{The threshold to define correlated rules.}
}
\value{
The rules matrix only with the rules selected
}
\description{
Discard Decision Rules too correlated (i.e. Cov(rule1,rule2)>t_corr).
}
\keyword{internal}
