% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_extreme_rules.R
\name{filter_extreme_rules}
\alias{filter_extreme_rules}
\title{Filter Extreme Decision Rules}
\usage{
filter_extreme_rules(rules_matrix, rules_list, t)
}
\arguments{
\item{rules_matrix}{The causal rules matrix.}

\item{rules_list}{The list of the causal rules.}

\item{t}{The threshold to define too generic or too specific rules.}
}
\value{
The rules matrix only with the rules selected
}
\description{
Discard rules with too few or too many observations.
}
\keyword{internal}
