% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cre.nls.R
\name{fit.cre.nls}
\alias{fit.cre.nls}
\title{Fit Constant Rate Effect Models (Nonlinear Least Squares)}
\usage{
fit.cre.nls(data, format, contexts = NULL, model, budget = 100,
  warnOnly = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{format}{One of \code{"wide"} or \code{"long"}; describes the format of the data frame.}

\item{contexts}{A vector of names of contexts to use in fitting. By default (\code{NULL}), all contexts available in the data are used.}

\item{model}{Model to fit: currently, one of \code{"logistic"}, \code{"bias"} or \code{"VRE"} (see Details).}

\item{budget}{Computational budget for fitting. Increasing the budget will improve goodness of fit but leads to longer runtimes. The default value of \code{100} is suitable in many cases, but the user is encouraged to experiment with the value.}

\item{warnOnly}{Whether a warning only should be issued when fitting fails. If \code{FALSE}, the routine exits with an error upon failure to fit.}
}
\value{
Object of class \code{"logistic"}, \code{"bias"} or \code{"VRE"}, depending on the \code{model} specified; a list with the following elements:
\describe{
\item{\code{data}}{Data the model was fit on, in long format}
\item{\code{parameters}}{Model parameters}
\item{\code{cumul_objfun_value}}{Sum of squared residuals between model and data}
\item{\code{cumul_objfun_value_normalized}}{Sum of squared residuals divided by number of non-\code{NA} data points}
\item{\code{N}}{Number of non-\code{NA} data points}
}
}
\description{
Fit Constant (and Variable) Rate Effect models to data using a nonlinear least squares algorithm.
}
\details{
It is possible to fit three kinds of models, controlled by the \code{model} argument: \code{"logistic"}, the classical model of a family of logistic curves with identical slopes but potentially varying intercepts; \code{"bias"}, the Kauhanen-Walkden production bias model; \code{"VRE"}, a family of logistic curves with independent slopes and intercepts.
}
