% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentize.R
\name{frequentize}
\alias{frequentize}
\title{From Response to Frequency Data}
\usage{
frequentize(data)
}
\arguments{
\item{data}{Data frame; must contain columns labelled \code{"date"}, \code{"context"} and \code{"response"}.}
}
\value{
A data frame.
}
\description{
Turn a response-based dataset into a frequency dataset.
}
