% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_causal_rules.R
\name{generate_causal_rules}
\alias{generate_causal_rules}
\title{Generate causal rules}
\usage{
generate_causal_rules(X, ite, method_params, hyper_params)
}
\arguments{
\item{X}{A covariate matrix.}

\item{ite}{An estimated ITE.}

\item{method_params}{A vector of method parameters.}

\item{hyper_params}{A vector of hyper parameters.}
}
\value{
A vector of causal rules.
}
\description{
Generates causal decision rules.
}
\keyword{internal}
