% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cre_dataset.R
\name{generate_cre_dataset}
\alias{generate_cre_dataset}
\title{Generate CRE Synthetic Data}
\usage{
generate_cre_dataset(
  n = 1000,
  rho = 0,
  n_rules = 2,
  p = 10,
  effect_size = 2,
  binary = FALSE
)
}
\arguments{
\item{n}{the number of observations}

\item{rho}{the correlation within the covariates (default: 0)}

\item{n_rules}{the number of causal rules, either 2 (default) or 4}

\item{p}{the number of covariates (default: 10)}

\item{effect_size}{the effect size magnitude (default: 0.5)
TODO: what is the range of effect size magnitude?}

\item{binary}{whether to use binary or continuous outcomes,
either TRUE (default) or FALSE}
}
\value{
a list of synthetic data containing an outcome vector, a treatment vector,
and a matrix of p covariates
}
\description{
Generates synthetic data with continues or binary outcome for testing the
Causal Rule Ensemble package
}
\examples{
set.seed(123)
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

}
