% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Generate Rules}
\usage{
generate_rules(
  X,
  ite_std,
  ntrees_rf,
  ntrees_gbm,
  node_size,
  max_nodes,
  max_depth,
  max_decay,
  type_decay,
  random_state
)
}
\arguments{
\item{X}{The covariate matrix.}

\item{ite_std}{The standardized ITE.}

\item{ntrees_rf}{The number of decision trees for randomForest.}

\item{ntrees_gbm}{The number of decision trees for gradient boosting.}

\item{node_size}{The minimum size of the trees' terminal nodes.}

\item{max_nodes}{The maximum number of terminal nodes trees in the forest can have.}

\item{max_depth}{The number of top levels from each tree considered
to extract conditions.}

\item{max_decay}{Decay Threshold for pruning the rules.}

\item{type_decay}{Decay Type for pruning the rules (1: relative error; 2: error).}

\item{random_state}{An integer number that represents a random state.}
}
\value{
A vector of causal rules.
}
\description{
Method for generating (Causal) Decision Rules before Regularization.
}
