% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules_matrix.R
\name{generate_ruls_matrix}
\alias{generate_ruls_matrix}
\title{Title}
\usage{
generate_ruls_matrix(rules_list, X)
}
\arguments{
\item{rules_list}{input variable description}

\item{X}{input variable description}
}
\value{

}
\description{
TBD
}
\examples{
TBD
}
