% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honest_splitting.R
\name{honest_splitting}
\alias{honest_splitting}
\title{Honest Splitting}
\usage{
honest_splitting(y, z, X, ratio_dis)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The features matrix.}

\item{ratio_dis}{the ratio of data delegated to the discovery subsample}
}
\value{
A list containing the discovery and inference subsamples
}
\description{
Splits data into discovery and inference subsamples.
}
\keyword{internal}
