% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_rules.R
\name{interpret_rules}
\alias{interpret_rules}
\title{Interpret rules}
\usage{
interpret_rules(rules, X_names)
}
\arguments{
\item{rules}{A vector of rules.}

\item{X_names}{A vector of real names of the covariates.}
}
\value{
A list of explicit (human-readable) rules.
}
\description{
Replaces the column numbers in the rules vector with their real names.
}
\keyword{internal}
