% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_select_rules.R
\name{interpret_select_rules}
\alias{interpret_select_rules}
\title{Interpret select rules}
\usage{
interpret_select_rules(select_rules_dis, X_names)
}
\arguments{
\item{select_rules_dis}{A vector of select causal rules.}

\item{X_names}{A vector of real names of the covariates.}
}
\value{
A vector of select causal rules that are interpretable.
}
\description{
Replaces the column numbers in the select rules vector with their real names.
}
\keyword{internal}
