% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cre.R
\name{plot_cre}
\alias{plot_cre}
\title{The Causal Rule Ensemble Output Plot}
\usage{
plot_cre(cre_results)
}
\arguments{
\item{cre_results}{the S3 object returned from the cre() function}
}
\value{
Returns nothing; plots the CRE results using ggplot
}
\description{
Plots the results of the CRE function: select causal rules and CATE estimates
}
\examples{
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

cre_results <- cre(y = dataset[["y"]], z = dataset[["z"]],
                   X = as.data.frame(dataset[["X"]]), ratio_dis = 0.25,
                   ite_method_dis = "bart", include_ps_dis = TRUE,
                   ite_method_inf = "bart", include_ps_inf = TRUE,
                   ntrees_rf = 100, ntrees_gbm = 50, min_nodes = 20,
                   max_nodes = 5, t = 0.025, q = 0.8)

plot_cre(cre_results)

}
