% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rule_frequency.R
\name{plot_rule_frequency}
\alias{plot_rule_frequency}
\title{Plot frequency of rule discovery}
\usage{
plot_rule_frequency(object, max_rules = 10)
}
\arguments{
\item{object}{A \code{cre} object.}

\item{max_rules}{Maximum number of decision rules to plot.}
}
\value{
A plot showing the frequency of rule discovery.
}
\description{
Generates a plot showing the number of times each rule (before filtering)
was discovered.
}
