% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cre}
\alias{predict.cre}
\title{Predict Individual Treatment Effect via Causal Rule Ensemble}
\usage{
\method{predict}{cre}(CRE, X)
}
\arguments{
\item{CRE}{A cre object from running the CRE function.}

\item{X}{A covariate matrix (or data.frame)}
}
\value{
Array with the estimated Individual Treatment Effects
}
\description{
Predict Individual Treatment Effect via Causal Rule Ensemble algorithm.
}
