% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data}
\usage{
prepare_data(data, format, contexts = NULL)
}
\arguments{
\item{data}{A data frame (e.g. from \code{read.csv})}

\item{format}{Format of the data frame: either "wide" or "long"}

\item{contexts}{Which contexts to use in fitting. By default (\code{NULL}), all contexts are used.}
}
\value{
A data frame in a format \code{\link{fit.cre.nls}} understands.
}
\description{
Prepare data for use with the curve-fitting routines in \code{\link{fit.cre.nls}}.
}
\details{
Under normal circumstances, there is no need for the end user to call this function directly: it is called automatically by the \code{\link{fit.cre.nls}} routine. User-level access is provided for completeness and for debugging purposes.
}
