% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cre}
\alias{print.cre}
\title{Extend print function for the CRE object}
\usage{
\method{print}{cre}(x, verbose = 2, ...)
}
\arguments{
\item{x}{A cre object from running the CRE function.}

\item{verbose}{Set level of results description details: 0 for only results
summary, 1 for results and parameters summary, 2 for results and parameters
and rules summary (default 2).}

\item{...}{Additional arguments passed to customize the results description.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Prints a brief summary of the CRE object
}
