% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_rules.R
\name{prune_rules}
\alias{prune_rules}
\title{Prune (Causal) Decision Rules}
\usage{
prune_rules(rules, X, ite_std, max_decay, type_decay)
}
\arguments{
\item{rules}{A list of decision rules.}

\item{X}{The features matrix.}

\item{ite_std}{The standardized ITE.}

\item{max_decay}{Decay Threshold for pruning the rules.}

\item{type_decay}{Decay Type for pruning the rules (1: relative error; 2: error).}
}
\value{
A vector of (Causal) Decision Rules.
}
\description{
Method for pruning (Causal) Decision Rules extracted from the Random Forest or
Gradient Boosting algorithms.
}
\keyword{internal}
