% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Title}
\usage{
split_data(y, z, X, ratio_dis)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{raio_dis}{the ratio of data delegated to the discovery subsample}
}
\value{
a list containing the discovery and inference subsamples
}
\description{
description
}
\examples{
TBD

}
