% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules_matrix.R
\name{standardize_rules_matrix}
\alias{standardize_rules_matrix}
\title{Standardize Rules Matrix}
\usage{
standardize_rules_matrix(rules_matrix)
}
\arguments{
\item{rules_matrix}{The rules matrix.}
}
\value{
Standardized rules matrix
}
\description{
Standardize (i.e. mean=0 and stdev=1) the rules matrix.
}
\keyword{internal}
