% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cre}
\alias{summary.cre}
\title{Print summary of CRE object}
\usage{
\method{summary}{cre}(object, verbose = 2)
}
\arguments{
\item{object}{A cre object from running the CRE function.}

\item{verbose}{Set level of results description details: only results summary
0, results+parameters summary 1, results+parameters+rules summary (default 2).}
}
\value{
A summary of the CRE object
}
\description{
Prints a brief summary of the CRE object
}
