% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sensitivity.R
\name{analyze_sensitivity}
\alias{analyze_sensitivity}
\title{Sensitivity Analysis}
\usage{
analyze_sensitivity(ite_std, rules_matrix_std)
}
\arguments{
\item{ite_std}{the standardized ITE}

\item{rules_matrix_std}{the standardized causal rules matrix}
}
\value{
a list containing the results of the sensitivity analysis
}
\description{
Method for analyzing the sensitivity of the estimates of the causal
rule-specific treatment effects
}
