% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cre}
\alias{autoplot.cre}
\title{A helper function for CRE object}
\usage{
\method{autoplot}{cre}(cre_results, method_params, ...)
}
\arguments{
\item{cre_results}{A CRE object.}

\item{method_params}{Method parameters.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot CRE object using ggplot2 package.
}
\keyword{internal}
