% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hyper_params.R
\name{check_hyper_params}
\alias{check_hyper_params}
\title{Check input parameters}
\usage{
check_hyper_params(params)
}
\arguments{
\item{params}{The list of parameters required to run the function.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Checks input parameters for the cre function.
}
\keyword{internal}
