% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hyper_params.R
\name{check_hyper_params}
\alias{check_hyper_params}
\title{Check input parameters}
\usage{
check_hyper_params(params)
}
\arguments{
\item{params}{The list of parameters required to run the function.}

\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The covariate matrix.}
}
\value{
A list of params that might be changed during checks.
}
\description{
Checks input parameters for the cre function.
}
