% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input parameters}
\usage{
check_input(y, z, X, params)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The covariate matrix.}

\item{params}{The list of parameters required to run the function.}
}
\value{
A list of params that might be changed during checks.
}
\description{
Checks input parameters for the cre function.
}
