% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{Check input data}
\usage{
check_input_data(y, z, X, ite = NULL)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The features matrix.}

\item{ite}{The estimated ITE vector.}
}
\value{
Number of data samples.
}
\description{
Conducts sanity checks for the input data
}
\keyword{internal}
