% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_input_data}
\alias{check_input_data}
\title{Check Input Data}
\usage{
check_input_data(y, z, X)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}
}
\value{
Number of data samples.
}
\description{
Conducts sanity checks for the input data
}
