% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre.R
\name{cre}
\alias{cre}
\title{Title}
\usage{
cre(y, z, X, ratio_dis, method_dis, method_inf, ntrees, min_nods, max_nodes)
}
\arguments{
\item{y}{input variable description}

\item{z}{input variable description}

\item{X}{input variable description}

\item{ratio_dis}{input variable description}

\item{method_dis}{input variable description}

\item{method_inf}{input variable description}

\item{ntrees}{input variable description}

\item{min_nods}{input variable description}

\item{max_nodes}{input variable description}
}
\value{

}
\description{
description
}
