% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre.R
\name{cre}
\alias{cre}
\title{The Causal Rule Ensemble}
\usage{
cre(y, z, X, params)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The covariate matrix.}

\item{params}{the list of parameters required to run the function, including:
\itemize{
\item \emph{Parameters for Discovery}
\itemize{
\item \emph{ratio_dis}: The ratio of data delegated to the discovery sub-sample.
\item \emph{ite_method_dis}: The method to estimate the discovery sample ITE.
\item \emph{include_ps_dis}: Whether or not to include propensity score estimate
as a covariate in discovery ITE estimation, considered only for BART, XBART,
or CF.
\item \emph{ps_method_dis}: The estimation model for the propensity score on the
discovery subsample.
\item \emph{or_method_dis}: The estimation model for the outcome regressions
estimate_ite_aipw on the discovery subsample.
}
\item \emph{Parameters for Inference}
\itemize{
\item \emph{ite_method_inf}: the method to estimate the inference sample ITE.
\item \emph{include_ps_inf} whether or not to include propensity score estimate as
a covariate in inference ITE estimation, considered only for BART, XBART,
or CF.
\item \emph{include_ps_inf}: whether or not to include propensity score estimate as
a covariate in inference ITE estimation, considered only for BART, XBART,
or CF.
\item \emph{ps_method_inf}: the estimation model for the propensity score on the
inference subsample.
\item \emph{or_method_inf}: the estimation model for the outcome regressions in
estimate_ite_aipw on the inference subsample.
}
\item \emph{Other Parameters}
\itemize{
\item \emph{ntrees_rf}: the number of decision trees for randomForest.
\item \emph{ntrees_gbm}: the number of decision trees for gradient boosting.
\item \emph{min_nodes}: the minimum size of the trees' terminal nodes.
\item \emph{max_nodes}: the maximum size of the trees' terminal nodes.
\item \emph{t}: the common support used in generating the causal rules matrix.
\item \emph{q}: the selection threshold used in selecting the causal rules.
\item \emph{rules_method}: the method for selecting causal rules with binary outcomes.
\item \emph{include_offset}: whether or not to include an offset when estimating
the ITE, for poisson only.
\item \emph{offset_name}: the name of the offset, if it is to be included.
\item \emph{cate_method}: the method to estimate the CATE values.
\item \emph{cate_SL_library}: the library used if cate_method is set to DRLearner.
\item \emph{filter_cate}: whether or not to filter rules with p-value <= 0.05.
}
}}
}
\value{
an S3 object containing the matrix of Conditional
Average Treatment Effect estimates
}
\description{
Performs the Causal Rule Ensemble on a dataset with a response variable,
a treatment variable, and various features
}
\examples{
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

cre_results <- cre(y = dataset[["y"]], z = dataset[["z"]],
                   X = as.data.frame(dataset[["X"]]),
                   params = list(ratio_dis = 0.25,
                                 ite_method_dis="brat",
                                 include_ps_dis = TRUE,
                                 ite_method_inf = "bart",
                                 include_ps_inf = TRUE,
                                 ntrees_rf = 100,
                                 ntrees_gbm = 50,
                                 min_nodes = 20,
                                 max_nodes = 5,
                                 t = 0.025,
                                 q = 0.8))

}
