% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre.R
\name{cre}
\alias{cre}
\title{The Causal Rule Ensemble}
\usage{
cre(
  y,
  z,
  X,
  ratio_dis,
  ite_method_dis,
  ite_method_inf,
  include_ps_dis = NA,
  include_ps_inf = NA,
  ntrees_rf,
  ntrees_gbm,
  min_nodes,
  max_nodes,
  t,
  q,
  rules_method
)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{ratio_dis}{the ratio of data delegated to the discovery subsample}

\item{ite_method_dis}{the method to estimate the discovery sample ITE}

\item{ite_method_inf}{the method to estimate the inference sample ITE}

\item{include_ps_dis}{whether or not to include propensity score estimate as a covariate in discovery ITE estimation, considered only for BART, XBART, or CF}

\item{include_ps_inf}{whether or not to include propensity score estimate as a covariate in inference ITE estimation, considered only for BART, XBART, or CF}

\item{ntrees_rf}{the number of decision trees for randomForest}

\item{ntrees_gbm}{the number of decision trees for gradient boosting}

\item{min_nodes}{the minimum size of the trees' terminal nodes}

\item{max_nodes}{the maximum size of the trees' terminal nodes}

\item{t}{the common support used in generating the causal rules matrix}

\item{q}{the selection threshold used in selecting the causal rules}

\item{rules_method}{the method for selecting causal rules with binary outcomes}
}
\value{
a list containing the list of select causal rules and a matrix of Conditional Average Treatment Effect estimates
}
\description{
Method for performing the Causal Rule Ensemble on a dataset with a response variable, a treatment variable, and various features
}
