% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_causal_rules.R
\name{discover_causal_rules}
\alias{discover_causal_rules}
\title{Discover Causal Decision Rules}
\usage{
discover_causal_rules(
  rules_matrix,
  rules_list,
  ite,
  stability_selection,
  cutoff,
  pfer,
  penalty_rl
)
}
\arguments{
\item{rules_matrix}{The causal rules matrix.}

\item{rules_list}{A vector of causal rules.}

\item{ite}{The estimated ITE.}

\item{stability_selection}{Whether or not using stability selection to
select the causal rules.}

\item{cutoff}{Threshold (percentage) defining the minimum cutoff value for
the stability scores.}

\item{pfer}{Upper bound for the per-family error rate (tolerated amount of
falsely selected rules).}

\item{penalty_rl}{Order of penalty for rules length during LASSO for Causal
Rules Discovery (i.e. 0: no penalty, 1: rules_length, 2: rules_length^2)}
}
\value{
List of the Causal Decision Rules discovered
}
\description{
Discovers Causal Decision Rules by randomization-based tests
}
\keyword{internal}
