% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_rules.R
\name{discover_rules}
\alias{discover_rules}
\title{Discover rules}
\usage{
discover_rules(X, ite, method_params, hyper_params)
}
\arguments{
\item{X}{A covariate matrix.}

\item{ite}{An estimated ITE.}

\item{method_params}{A vector of method parameters.}

\item{hyper_params}{A vector of hyper parameters.}
}
\value{
A minimal set of rules linearly decomposing the Conditional Average
Treatment Effect (CATE).
}
\description{
Discover the minimal set of rules linearly decomposing the Conditional
Average Treatment Effect (CATE).
}
\keyword{internal}
