% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{Estimate the Conditional Average Treatment Effect (CATE)}
\usage{
estimate_cate(
  y_inf,
  z_inf,
  X_inf,
  X_names,
  offset,
  rules_matrix_inf,
  select_rules_interpretable,
  cate_method,
  ite_inf,
  sd_ite_inf,
  cate_SL_library,
  t_pvalue
)
}
\arguments{
\item{y_inf}{The outcome vector for the inference subsample.}

\item{z_inf}{The treatment vector for the inference subsample.}

\item{X_inf}{The covariate vector for the inference subsample.}

\item{X_names}{The names of the covariates.}

\item{offset}{Name of the covariate to use as offset (i.e. 'x1') for
Poisson Estimation. NULL if offset is not used.}

\item{rules_matrix_inf}{The standardized causal rules matrix for the
inference subsample.}

\item{select_rules_interpretable}{The list of select causal rules in terms of
coviariate names.}

\item{cate_method}{The method to estimate the CATE values.}

\item{ite_inf}{The estimated ITEs for the inference subsample.}

\item{sd_ite_inf}{The standard deviations for the estimated ITEs for the
inference subsample.}

\item{cate_SL_library}{The library used if cate_method is DRLearner.}

\item{t_pvalue}{The threshold to define statistically significant rules
(filter only causal decision rules with p-value <= t_pvalue).}
}
\value{
a matrix of CATE estimates
}
\description{
Estimates the Conditional Average Treatment Effect given a standardized
vector of Individual Treatment Effects, a standardized matrix of causal
rules, a list of causal rules.
}
\keyword{internal}
