% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{Estimate the Conditional Average Treatment Effect}
\usage{
estimate_cate(
  y_inf,
  z_inf,
  X_inf,
  rules_matrix_inf,
  rules_explicit,
  ite_inf,
  t_pvalue
)
}
\arguments{
\item{y_inf}{The outcome vector for the inference subsample.}

\item{z_inf}{The treatment vector for the inference subsample.}

\item{X_inf}{The covariate vector for the inference subsample.}

\item{rules_matrix_inf}{The rules matrix for the
inference subsample.}

\item{rules_explicit}{The list of select rules in terms of
coviariate names (explicit).}

\item{ite_inf}{The estimated ITEs for the inference subsample.}

\item{t_pvalue}{The threshold to define statistically significant rules
(filter only rules with p-value <= t_pvalue).}
}
\value{
A dataframe summarizing the CATE linear decomposition:
\itemize{
\item 'Rule': rule name,
\item 'Estimate': linear contribution to CATE,
\item 'CI_lower`: lower bound 95\% confidence interval on the estimate,
\item 'CI_upper`: upper bound 95\% confidence interval on the estimate,
\item \code{P-Value}: p-value (from Z-test).
}
}
\description{
Estimates the Conditional Average Treatment Effect (CATE) by
linearly modeling the Individual Treatment Effect on the Inference subsample
via the rules selected in the Discovery step.
}
\keyword{internal}
