% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite.R
\name{estimate_ite}
\alias{estimate_ite}
\title{Estimate the Individual Treatment Effect}
\usage{
estimate_ite(
  y,
  z,
  X,
  ite_method,
  include_ps,
  binary,
  X_names,
  include_offset,
  offset_name
)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{ite_method}{the method for estimating the Individual Treatment Effect, either Inverse Propensity Weighting "ipw", Stabilized Inverse Propensity Weighting "sipw", Outcome Regression "or", BART "bart", Accelerated BART "xbart", Bayesian Causal Forest "bcf", Accelerated Bayesian Causal Forest "xbcf", or Causal Forest "cf"}

\item{include_ps}{whether or not to include propensity score estimate as a covariate in ITE estimation}

\item{binary}{whether or not the outcome is binary}

\item{X_names}{the names of the covariates}

\item{include_offset}{whether or not to include an offset when estimating the ITE, for poisson only}

\item{offset_name}{the name of the offset, if it is to be included}
}
\value{
a list of raw ITE estimates, standardized ITE estimates, and standard deviations for the ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect given a response vector, a treatment vector, a features matrix, and a desired algorithm
}
