% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite.R
\name{estimate_ite}
\alias{estimate_ite}
\title{Title}
\usage{
estimate_ite(y, z, X, method)
}
\arguments{
\item{y}{input variable description}

\item{z}{input variable description}

\item{X}{input variable description}

\item{method}{input variable description}
}
\value{

}
\description{
description
}
\examples{
TBD
}
