% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_bart.R
\name{estimate_ite_bart}
\alias{estimate_ite_bart}
\title{Estimate the Individual Treatment Effect using Bayesian Additive Regression Trees}
\usage{
estimate_ite_bart(y, z, X, include_ps)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{include_ps}{whether or not to include propensity score estimate as a covariate in ITE estimation}
}
\value{
a vector of ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Bayesian Additive Regression Trees given a response vector, a treatment vector, and a features matrix
}
\examples{
#TBD

}
