% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_bart.R
\name{estimate_ite_bart}
\alias{estimate_ite_bart}
\title{Estimate the Individual Treatment Effect (ITE) using Bayesian Additive
Regression Trees (BART)}
\usage{
estimate_ite_bart(y, z, X, learner_ps)
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{learner_ps}{Method for the estimation of the propensity score.}
}
\value{
A list of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using Bayesian Additive Regression
Trees given a response vector, a treatment vector, and a features matrix.
}
\note{
The number of samples and the number of burn are set by default equal
to 500.
}
\keyword{internal}
