% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estiamte_ite_bcf.R
\name{estimate_ite_bcf}
\alias{estimate_ite_bcf}
\title{Title}
\usage{
estimate_ite_bcf(y, z, X)
}
\arguments{
\item{y}{input variable description}

\item{z}{input variable description}

\item{X}{input variable description}
}
\value{

}
\description{
description
}
\examples{
TBD
}
