% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_bcf.R
\name{estimate_ite_bcf}
\alias{estimate_ite_bcf}
\title{Estimate the Individual Treatment Effect (ITE) using Bayesian Causal Forest
(BCF)}
\usage{
estimate_ite_bcf(y, z, X, ps_method)
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{ps_method}{A method for the estimation of the propensity score.}
}
\value{
A list of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using Bayesian Causal Forest given
a response vector, a treatment vector, and a features matrix.
}
\keyword{internal}
