% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_bcf.R
\name{estimate_ite_bcf}
\alias{estimate_ite_bcf}
\title{Estimate the Individual Treatment Effect using Bayesian Causal Forest}
\usage{
estimate_ite_bcf(y, z, X)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}
}
\value{
a vector of ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Bayesian Causal Forest given a response vector, a treatment vector, and a features matrix
}
