% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_cf.R
\name{estimate_ite_cf}
\alias{estimate_ite_cf}
\title{Estimate the Individual Treatment Effect using Causal Forest}
\usage{
estimate_ite_cf(y, z, X, include_ps)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{include_ps}{whether or not to include propensity score estimate as a covariate in ITE estimation}
}
\value{
a list of ITE estimates and standard deviations for the ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Causal Forest given a response vector, a treatment vector, and a features matrix
}
