% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_ipw.R
\name{estimate_ite_ipw}
\alias{estimate_ite_ipw}
\title{Estimate the Individual Treatment Effect using Inverse Propensity Weighting}
\usage{
estimate_ite_ipw(y, z, X)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}
}
\value{
a vector of ITE estimates
}
\description{
Estimates the Individual Treatment Effect using Inverse Propensity Weighting
given a response vector, a treatment vector, and a features matrix.
}
