% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_or.R
\name{estimate_ite_or}
\alias{estimate_ite_or}
\title{Estimate the Individual Treatment Effect using Outcome Regression}
\usage{
estimate_ite_or(y, z, X)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}
}
\value{
a vector of ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Outcome Regression given a response vector, a treatment vector, and a features matrix
}
\examples{
TBD

}
