% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_oreg.R
\name{estimate_ite_oreg}
\alias{estimate_ite_oreg}
\title{Estimate the Individual Treatment Effect using Outcome Regression}
\usage{
estimate_ite_oreg(y, z, X)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}
}
\value{
a vector of ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Outcome
Regression given a response vector, a treatment vector, and a features matrix.
}
\examples{
dataset <- generate_cre_dataset(n = 500, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- dataset[["y"]]
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])

ite_list <- estimate_ite_oreg(y, z, X)

}
