% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_poisson.R
\name{estimate_ite_poisson}
\alias{estimate_ite_poisson}
\title{Estimate the Individual Treatment Effect using Poisson Regression}
\usage{
estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{X_names}{the names of the covariates}

\item{include_offset}{whether to include an offset (i.e. model outcome rate)
or not (i.e. model outcome counts)}

\item{offset_name}{the name of the covariate to use as offset (i.e. 'x1')}
}
\value{
a vector of ITE estimates
}
\description{
Estimates the Individual Treatment Effect using Poisson Regression given a
response vector, a treatment vector, and a features matrix.
}
\keyword{internal}
