% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_poisson.R
\name{estimate_ite_poisson}
\alias{estimate_ite_poisson}
\title{Estimate the Individual Treatment Effect using Poisson Regression}
\usage{
estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{X_names}{the names of the covariates}

\item{include_offset}{whether or not to include an offset when estimating the ITE, for poisson only}

\item{offset_name}{the name of the offset, if it is to be included}
}
\value{
a vector of ITE estimates
}
\description{
Method for estimating the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix
}
