% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_poisson.R
\name{estimate_ite_poisson}
\alias{estimate_ite_poisson}
\title{Estimate the Individual Treatment Effect using Poisson Regression}
\usage{
estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
}
\arguments{
\item{y}{the observed response vector}

\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{X_names}{the names of the covariates}

\item{include_offset}{whether or not to include an offset when estimating
the ITE, for poisson only}

\item{offset_name}{the name of the offset, if it is to be included}
}
\value{
a vector of ITE estimates
}
\description{
Estimates the Individual Treatment Effect using Poisson Regression given a
response vector, a treatment vector, and a features matrix.
}
\examples{
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

}
