% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_sipw.R
\name{estimate_ite_sipw}
\alias{estimate_ite_sipw}
\title{Estimate the Individual Treatment Effect (ITE) using Stabilized Inverse
Probability Weighting (SIPW)}
\usage{
estimate_ite_sipw(y, z, X, ps_method)
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{ps_method}{An estimation method for the propensity score.}
}
\value{
A vector of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using Stabilized Inverse
Probability Weighting given a response vector, a treatment vector, and a
features matrix.
}
\keyword{internal}
