% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_tlearner.R
\name{estimate_ite_tlearner}
\alias{estimate_ite_tlearner}
\title{Estimate the Individual Treatment Effect (ITE) using T-Learner}
\usage{
estimate_ite_tlearner(y, z, X, learner_y = "SL.xgboost")
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{learner_y}{A estimation model for the outcome.}
}
\value{
A list of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using T-Learner given a response
vector, a treatment vector, a features matrix and estimation model for the
outcome.
}
\keyword{internal}
