% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ite_xlearner.R
\name{estimate_ite_xlearner}
\alias{estimate_ite_xlearner}
\title{Estimate the Individual Treatment Effect (ITE) using X-Learner.}
\usage{
estimate_ite_xlearner(y, z, X, learner_y = "SL.xgboost")
}
\arguments{
\item{y}{An observed response vector.}

\item{z}{A treatment vector.}

\item{X}{A features matrix.}

\item{learner_y}{A estimation model for the outcome.}
}
\value{
A list of ITE estimates.
}
\description{
Estimates the Individual Treatment Effect using X-Learner given a response
vector, a treatment vector, a features matrix and an estimation model for
the outcome.
}
\keyword{internal}
