% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ps.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimate the propensity score}
\usage{
estimate_ps(z, X, ps_method = "SL.xgboost")
}
\arguments{
\item{z}{A treatment vector.}

\item{X}{A features data frame.}

\item{ps_method}{An estimation model for the propensity score
(default: \code{SL.xgboost}).}
}
\value{
A vector of propensity score estimates.
}
\description{
Estimates the Propensity Score given a treatment vector and
features data frame.
}
\keyword{internal}
