% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ps.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimate the Propensity Score}
\usage{
estimate_ps(z, X, ps_method = "SL.xgboost")
}
\arguments{
\item{z}{the treatment vector}

\item{X}{the features matrix}

\item{ps_method}{the estimation model for the propensity score}
}
\value{
a list of propensity score estimates
}
\description{
Method for estimating the Propensity Score given a treatment vector and features matrix
}
\examples{
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
ps_method <- "SL.xgboost"

est_ps <- estimate_ps(z, X, ps_method)

}
