% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ps.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimate the Propensity Score}
\usage{
estimate_ps(z, X)
}
\arguments{
\item{z}{the treatment vector}

\item{X}{the features matrix}
}
\value{
a list of propensity score estimates
}
\description{
Method for estimating the Propensity Score given a treatment vector and features matrix
}
