% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate}
\usage{
evaluate(ground_truth, prediction)
}
\arguments{
\item{ground_truth}{List of true (CDR/EM) values.}

\item{prediction}{List of predicted (CDR/EM) values.}
}
\value{
Intersection Over Union, Precision, Recall
}
\description{
Compute different metrics for discovery evaluation
}
\keyword{internal}
