% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effect_modifiers.R
\name{extract_effect_modifiers}
\alias{extract_effect_modifiers}
\title{Extract Effect Modifiers}
\usage{
extract_effect_modifiers(rules_list, X_names)
}
\arguments{
\item{rules_list}{A list of (Causal) Decision Rules}

\item{X_names}{A list of the Covariate names.}
}
\value{
List of the Effect Modifiers.
}
\description{
Method for extracting the Effect Modifiers from a list of (Causal) Decision
Rules.
}
\keyword{internal}
