% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rules.R
\name{extract_rules}
\alias{extract_rules}
\title{Extract Causal Rules}
\usage{
extract_rules(treelist, X, ntrees, ite_std, take_1, type_decay)
}
\arguments{
\item{treelist}{A list of decision trees.}

\item{X}{The features matrix.}

\item{ntrees}{The number of decision trees.}

\item{ite_std}{The standardized ITE.}

\item{take_1}{Whether or not to call the take1 helper function.}

\item{type_decay}{The type of decay to apply when pruning the rules.}
}
\value{
A vector of causal rules.
}
\description{
Method for extracting causal rules from the Random Forest or Gradient
Boosting algorithms.
}
\keyword{internal}
