% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rules.R
\name{extract_rules}
\alias{extract_rules}
\title{Extract (causal) decision rules}
\usage{
extract_rules(treelist, X, max_depth, digits = 2)
}
\arguments{
\item{treelist}{A list of decision trees.}

\item{X}{Features matrix.}

\item{max_depth}{A number of top levels from each tree considered.}

\item{digits}{A Number of digits for rounding decision rules to extract
conditions.}
}
\value{
A vector of (causal) decision rules.
}
\description{
Extracts causal rules from the random forest or the gradient
boosting algorithms.
}
\keyword{internal}
