% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rules.R
\name{extract_rules}
\alias{extract_rules}
\title{Extract (Causal) Decision Rules}
\usage{
extract_rules(treelist, X, ntrees, max_depth, digits = 2)
}
\arguments{
\item{treelist}{A list of decision trees.}

\item{X}{The features matrix.}

\item{ntrees}{The number of (the first) decision trees considered.}

\item{max_depth}{The number of top levels from each tree considered}

\item{digits}{The Number of digitits for rounding decision rules.
to extract conditions.}
}
\value{
A vector of (Causal) Decision Rules.
}
\description{
Method for extracting causal rules from the Random Forest or Gradient
Boosting algorithms.
}
\keyword{internal}
