% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_correlated_rules.R
\name{filter_correlated_rules}
\alias{filter_correlated_rules}
\title{Filter correlated rules}
\usage{
filter_correlated_rules(rules_matrix, rules_list, t_corr)
}
\arguments{
\item{rules_matrix}{A rules matrix.}

\item{rules_list}{A list of rules (names).}

\item{t_corr}{A threshold to define correlated rules.}
}
\value{
A rules matrix without the highly correlated rules (columns).
}
\description{
Discards highly correlated rules (i.e., \eqn{Cov(rule_1,rule_2) > t_{corr}}).
}
\keyword{internal}
