% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_irrelevant_rules.R
\name{filter_irrelevant_rules}
\alias{filter_irrelevant_rules}
\title{Filter Irrelevant Decision Rules}
\usage{
filter_irrelevant_rules(rules, X, ite_std, max_decay, type_decay)
}
\arguments{
\item{rules}{A list of decision rules.}

\item{X}{The features matrix.}

\item{ite_std}{The standardized ITE.}

\item{max_decay}{Decay Threshold for pruning the rules.}

\item{type_decay}{Decay Type for pruning the rules
(1: relative error; 2: error).}
}
\value{
A list of the selected Decision Rules.
}
\description{
Filter Irrelevant Decision Rules extracted evaluating the performance
reduction removing a specific rule at the time
(see 'Interpreting tree ensembles with inTrees' by Houtao Deng, 2019)
}
\keyword{internal}
