% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_irrelevant_rules.R
\name{filter_irrelevant_rules}
\alias{filter_irrelevant_rules}
\title{Filter irrelevant decision rules using leave-one-out pruning}
\usage{
filter_irrelevant_rules(rules, X, ite, t_decay)
}
\arguments{
\item{rules}{A list of rules.}

\item{X}{A features matrix.}

\item{ite}{An estimated ITE.}

\item{t_decay}{The decay threshold for rules pruning.}
}
\value{
A list of 'relevant' rules.
}
\description{
Filters the irrelevant decision rules. The irrelevant rules are interpreted
as an error increase after removing a variable-value pair from the decision rules.
(see "Interpreting tree ensembles with the inTrees package"
by Houtao Deng, 2019).
}
\keyword{internal}
