% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_causal_rules.R
\name{generate_causal_rules}
\alias{generate_causal_rules}
\title{Generate Causal Rules}
\usage{
generate_causal_rules(X, ite_std, method_params, hyper_params)
}
\arguments{
\item{X}{The covariate matrix.}

\item{ite_std}{The standardized ITE.}

\item{method_params}{Method Parameters.}

\item{hyper_params}{Hyper Parameters.}
}
\value{
A vector of Causal Rules.
}
\description{
Method for generating Causal Decision Rules.
}
\keyword{internal}
