% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cre_dataset.R
\name{generate_cre_dataset}
\alias{generate_cre_dataset}
\title{Generate CRE Synthetic Data}
\usage{
generate_cre_dataset(n, rho, n_rules, effect_size, binary)
}
\arguments{
\item{n}{the number of observations}

\item{rho}{the correlation within the variables}

\item{n_rules}{the number of causal rules, either 2 or 4}

\item{effect_size}{the effect size magnitude}

\item{binary}{whether to use binary or continuous outcomes, either TRUE or FALSE}
}
\value{
a list of synthetic data containing an outcome vector, a treatment vector,
and a matrix of 10 covariates
}
\description{
Generates synthetic data with continues or binary outcome for testing the
Causal Rule Ensemble package
}
\examples{
set.seed(123)
dataset_c <- generate_cre_dataset(n = 500, rho = 0, n_rules = 2,
effect_size = 0.5, binary = FALSE)

}
