% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cre_dataset.R
\name{generate_cre_dataset}
\alias{generate_cre_dataset}
\title{Generate CRE Synthetic Data}
\usage{
generate_cre_dataset(
  n = 1000,
  rho = 0,
  n_rules = 2,
  p = 10,
  effect_size = 2,
  binary_covariates = TRUE,
  binary_outcome = TRUE,
  confounding = "no"
)
}
\arguments{
\item{n}{An integer number that represents the number of observations.
Non-integer values will be converted into an integer number.}

\item{rho}{A positive double number (0,1) that represents the correlation
within the covariates (default: 0).}

\item{n_rules}{The number of causal rules, either 2 (default) or 4.}

\item{p}{The number of covariates (default: 10).}

\item{effect_size}{The effect size magnitude (default: 2).}

\item{binary_covariates}{Whether to use binary or continuous covariates
(default: TRUE)}

\item{binary_outcome}{Whether to use binary or continuous outcomes
(default: TRUE)}

\item{confounding}{Only for continuous outcome,add confounding variables:
linear confounding "lin", non-linear confounding "nonlin",
no confounding "no" (default: "no")}
}
\value{
A list of synthetic data containing:
\itemize{
\item the outcome vector (y),
\item the treatment vector (z),
\item the covariates matrix (X) and
\item the individual treatment vector (ite)
}
}
\description{
Generates synthetic data with continues or binary outcome
}
\note{
Set (binary/continuous) covariates domain (binary_covariates)
Set (binary/continuous) outcome domain (binary_outcome)
Increase complexity of the synthetic data set:
\itemize{
\item decreasing the sample size (n)
\item adding correlation among variables (rho)
\item increasing the number of rules (n_rules)
\item increasing the number of covariates (p)
\item decreasing the absolute value of the causal effect (effect_size)
\item adding linear or not-linear confounders (confounding)
}
}
\examples{
set.seed(123)
dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary_covariates = TRUE,
                                binary_outcome = TRUE, confounding = "no")

}
