% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Generate Rules}
\usage{
generate_rules(
  X,
  ite_std,
  ntrees_rf,
  ntrees_gbm,
  min_nodes,
  max_nodes,
  random_state
)
}
\arguments{
\item{X}{The covariate matrix.}

\item{ite_std}{The standardized ITE.}

\item{ntrees_rf}{The number of decision trees for randomForest.}

\item{ntrees_gbm}{The number of decision trees for gradient boosting.}

\item{min_nodes}{The minimum size of the trees' terminal nodes.}

\item{max_nodes}{The maximum size of the trees' terminal nodes.}

\item{random_state}{An integer number that repesents a random state.}
}
\value{
a vector of causal rules
}
\description{
Method for generating causal rules.
}
