% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Generate rules}
\usage{
generate_rules(
  X,
  ite,
  intervention_vars,
  ntrees_rf,
  ntrees_gbm,
  node_size,
  max_nodes,
  max_depth,
  replace
)
}
\arguments{
\item{X}{A covariate matrix.}

\item{ite}{A vector of estimated ITE.}

\item{intervention_vars}{A vector of intervention-able variables used for
rules generation.}

\item{ntrees_rf}{A number of decision trees for the random forest algorithm.}

\item{ntrees_gbm}{A number of decision trees for the gradient boosting
algorithm.}

\item{node_size}{A minimum size of the trees' terminal nodes.}

\item{max_nodes}{A maximum allowed number of terminal nodes per a tree in the
forest.}

\item{max_depth}{A number of top levels from each tree considered
to extract conditions.}

\item{replace}{A Boolean variable for replacement in bootstrapping.}
}
\value{
A list of generated decision rules
}
\description{
Generates a set of relevant decision rules characterizing the
heterogeneity in the ITE.
}
\keyword{internal}
