% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Title}
\usage{
generate_rules(tau, X, ntrees, min_nodes, max_nodes, method)
}
\arguments{
\item{tau}{input variable description}

\item{X}{input variable description}

\item{ntrees}{input variable description}

\item{min_nodes}{input variable description}

\item{max_nodes}{input variable description}

\item{method}{input variable description}
}
\value{

}
\description{
TBD
}
\examples{
TBD
}
