% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Generate Rules}
\usage{
generate_rules(X, ite_std, ntrees_rf, ntrees_gbm, min_nodes, max_nodes)
}
\arguments{
\item{X}{the features matrix}

\item{ite_std}{the standardized ITE}

\item{ntrees_rf}{the number of decision trees for randomForest}

\item{ntrees_gbm}{the number of decision trees for gradient boosting}

\item{min_nodes}{the minimum size of the trees' terminal nodes}

\item{max_nodes}{the maximum size of the trees' terminal nodes}
}
\value{
a vector of causal rules
}
\description{
Method for generating causal rules.
}
