% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Generate Rules}
\usage{
generate_rules(
  X,
  ite_std,
  intervention_vars,
  ntrees_rf,
  ntrees_gbm,
  node_size,
  max_nodes,
  max_depth,
  replace,
  random_state
)
}
\arguments{
\item{X}{The covariate matrix.}

\item{ite_std}{The standardized ITE.}

\item{intervention_vars}{Intervention-able variables used for Rules Generation.}

\item{ntrees_rf}{The number of decision trees for randomForest.}

\item{ntrees_gbm}{The number of decision trees for gradient boosting.}

\item{node_size}{The minimum size of the trees' terminal nodes.}

\item{max_nodes}{The maximum number of terminal nodes trees in the forest can have.}

\item{max_depth}{The number of top levels from each tree considered
to extract conditions.}

\item{replace}{Boolean variable for replacement in bootstrapping.}

\item{random_state}{An integer number that represents a random state.}
}
\value{
List of generated Decision Rules
}
\description{
Method for generating a set of relevant Decision Rules characterizing the
heterogeneity in the ITE
}
\keyword{internal}
