% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules.R
\name{generate_rules}
\alias{generate_rules}
\title{Generate rules}
\usage{
generate_rules(X, ite, ntrees, node_size, max_nodes, max_depth)
}
\arguments{
\item{X}{A covariate matrix.}

\item{ite}{A vector of estimated ITE.}

\item{ntrees}{The number of decision trees for the random forest algorithm.}

\item{node_size}{Minimum size of the trees' terminal nodes.}

\item{max_nodes}{Maximum number of terminal nodes per tree.}

\item{max_depth}{Maximum rules length.}
}
\value{
A list of rules (names).
}
\description{
Generates a list of rules characterizing the heterogeneity in the Conditional
Average Treatment Effect (CATE) by tree-based methods (i.e., random forest).
}
\keyword{internal}
