% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules_matrix.R
\name{generate_rules_matrix}
\alias{generate_rules_matrix}
\title{Generate Rules Matrix}
\usage{
generate_rules_matrix(X, rules_list, t)
}
\arguments{
\item{X}{the features matrix}

\item{rules_list}{a vector of causal rules}

\item{t}{the common support used in generating the causal rules matrix}
}
\value{
a list with a raw matrix of causal rules, a standardized matrix of causal rules, and a vector of causal rules
}
\description{
Method for generating a matrix of causal rules given a list
}
