% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules_matrix.R
\name{generate_rules_matrix}
\alias{generate_rules_matrix}
\title{Generate Rules Matrix}
\usage{
generate_rules_matrix(X, rules_list, t)
}
\arguments{
\item{X}{The features matrix.}

\item{rules_list}{A vector of causal rules.}

\item{t}{The common support used in generating the causal rules matrix.}
}
\value{
a list with:
\itemize{
\item a raw matrix of causal rules
\item a standardized matrix of causal rules, and
\item a vector of causal rules
}
}
\description{
Generates a matrix of causal rules given a list.
}
\keyword{internal}
