% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rules_matrix.R
\name{generate_rules_matrix}
\alias{generate_rules_matrix}
\title{Generate rules matrix}
\usage{
generate_rules_matrix(X, rules_list)
}
\arguments{
\item{X}{Features matrix.}

\item{rules_list}{A vector of causal rules.}
}
\value{
A causal rules matrix.
}
\description{
Generates a matrix of causal rules given a list.
}
\keyword{internal}
