% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_rules_filter.R
\name{lasso_rules_filter}
\alias{lasso_rules_filter}
\title{Filter (Causal) Decision Rules by LASSO (with stability selection)}
\usage{
lasso_rules_filter(
  rules_matrix_std,
  rules_list,
  ite_std,
  stability_selection,
  cutoff,
  pfer
)
}
\arguments{
\item{rules_matrix_std}{The standardized causal rules matrix.}

\item{rules_list}{A vector of causal rules.}

\item{ite_std}{The standardized ITE.}

\item{stability_selection}{Whether or not using stability selection to
select the causal rules.}

\item{cutoff}{Threshold defining the minimum cutoff value for the stability
scores.}

\item{pfer}{Upper bound for the per-family error rate (tolerated amount of
falsely selected rules).}
}
\value{
a vector of causal rules
}
\description{
Selects the causal rules that are most important.
}
\keyword{internal}
