% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cre}
\alias{plot.cre}
\title{Extend generic plot functions for cre class}
\usage{
\method{plot}{cre}(x, ...)
}
\arguments{
\item{x}{A cre object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot2 object, invisibly. This function is called for side effects.
}
\description{
A wrapper function to extend generic plot functions for cre class.
}
