% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cre}
\alias{print.cre}
\title{Extend print function for the CRE object}
\usage{
\method{print}{cre}(cre_results, method_params)
}
\arguments{
\item{cre_results}{A cre object from running the CRE function}

\item{method_params}{Method parameters.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Prints a brief summary of the CRE object
}
