% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cre}
\alias{print.cre}
\title{Extend print function for the CRE object}
\usage{
\method{print}{cre}(x, ...)
}
\arguments{
\item{x}{A cre object from running the CRE function}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Prints a brief description of the CRE object
}
