% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causal_rules.R
\name{select_causal_rules}
\alias{select_causal_rules}
\title{Title}
\usage{
select_causal_rules(tau, rules_matrix)
}
\arguments{
\item{tau}{input variable description}

\item{rules_matrix}{input variable description}
}
\value{

}
\description{
Description
}
\examples{
TBD
}
