% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causal_rules.R
\name{select_causal_rules}
\alias{select_causal_rules}
\title{Select Causal Rules}
\usage{
select_causal_rules(
  rules_matrix_std,
  rules_list,
  ite_std,
  binary,
  q,
  rules_method
)
}
\arguments{
\item{rules_matrix_std}{the standardized causal rules matrix}

\item{rules_list}{a vector of causal rules}

\item{ite_std}{the standardized ITE}

\item{binary}{whether or not the outcome is binary}

\item{q}{the selection threshold used in selecting the causal rules}

\item{rules_method}{the method for selecting causal rules with binary outcomes, either "conservative", "anticonservative", or NA}
}
\value{
a vector of causal rules
}
\description{
Method for selecting the causal rules that are most important
}
