% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causal_rules.R
\name{select_causal_rules}
\alias{select_causal_rules}
\title{Select Causal Rules}
\usage{
select_causal_rules(
  rules_matrix_std,
  rules_list,
  ite_std,
  q,
  stability_selection,
  pfer_val
)
}
\arguments{
\item{rules_matrix_std}{The standardized causal rules matrix.}

\item{rules_list}{A vector of causal rules.}

\item{ite_std}{The standardized ITE.}

\item{q}{The selection threshold used in selecting the causal rules.}

\item{stability_selection}{Whether or not using stability selection for selecting the causal rules.}

\item{pfer_val}{The Per-Family Error Rate, the expected number of false discoveries.}
}
\value{
a vector of causal rules
}
\description{
Selects the causal rules that are most important.
}
