% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Title}
\usage{
split_data(y, z, X, raio_dis)
}
\arguments{
\item{y}{variable description}

\item{z}{variable description}

\item{X}{variable description}

\item{raio_dis}{variable description}
}
\value{

}
\description{
description
}
\examples{
TBD
}
