% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(y, z, X, ratio_dis)
}
\arguments{
\item{y}{The observed response vector.}

\item{z}{The treatment vector.}

\item{X}{The features matrix.}

\item{ratio_dis}{the ratio of data delegated to the discovery subsample}
}
\value{
A list containing the discovery and inference subsamples
}
\description{
Splits data into discovery and inference subsamples.
}
\keyword{internal}
