% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cre}
\alias{summary.cre}
\title{Print summary of CRE object}
\usage{
\method{summary}{cre}(object, ...)
}
\arguments{
\item{object}{A cre object from running the CRE function}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
A summary of the CRE object
}
\description{
Prints a brief summary of the CRE object
}
